Add config option to enforce that tokens should have a token id for verification (and original id for refresh).
I think it's preferable to reject these, but for someone upgrading a running service it would be unsafe to start insisting that all tokens have token ids immediately. Having a setting allows the operator to start enforcing it after all of their valid tokens were made by a new enough version of the library, which depends on how they have configured expiration times.

(I need to update the docs once I know which version the jti change will ship in.)