Only include the orig_jti on refreshed tokens.
This means that someone who wants the original id for a non-refreshed token would need to read it from the 'jti' claim, but I think that's better than having quite so much redundant data in every original token.

(I'm tempted to do something similar for orig_iat, but I worry that would cause compatibility problems for anyone reading from the claim outside of the library. Also the timestamp is much shorter than a UUID, so it's a much smaller saving.)