Allow three states for the token id config.
I think this makes sense, since a UUID is a fairly hefty beast ad not everyone will need them. Thanks to @nigoroll for the suggestion.

This will make using token ids for invalidation a little trickier, but we can warn people that turning this off creates a limitation when we come to that or make them mandatory once we're making them more useful.