Handle more token errors for check_payload.
There are various subclasses of InvalidTokenError that end up bubbling up as unhandled errors in the old code, including InvalidAlgorithmError.

This is very similar to the fix in #72, but for the check_payload version of the similar logic.

Again, I've kept the explicit handling of DecodeError and Expired Signature, which are also children of InvalidTokenError, so that we don't change the behaviour for previously handled errors.

I think it would be good to unify the logic for these at some point, but for now I'm just handling the same errors in both places.