Store a token id for invalidated tokens, if we have them.
For refreshed tokens, I'm storing the original token's id, so that the blacklist lookup can match for any token from the same chain of refreshes. (The orig_jti claim of a refreshed token matches the jti claim of the original token, and is copied down the chain of refreshes. For original tokens, the orig_jti claim is absent, and we use the jti claim.)