Permit storing either the token or its original id (or both) for blacklist entries.
For Django 2.2 and above, also add constraint that at least one of the values is non-null. Since we allow configuring whether token ids are included, the token_id column needs to be optional. Where we can rely on token ids, I am going to storing the whole token value, so that also needs to be nullable.