Fix decode call for migration.
I don't care about verifying the token value here; if it got as far as the database, it's reasonable to expect it was valid when it was inserted.

This was my mistake in Styria-Digital/django-rest-framework-jwt#84, which was released in v1.18.0, where I didn't test the migration properly with real data.